Rename r53 healthcheck to bring under the 64 char limit
The aws_route53_health_check has a name limit of 64 chars, this limit
gets hit because terraform appends "-terraform-<datestamp>" to the end
of the name whilst it creates the resource for tracking, and after
the resource is created it removes the "-terraform-<datestamp>" text.

This sets a character size limit to an environment name of 9 chars when
using the naming format `${var.Env-Name}-${var.aws-region-name}-frontend-%d`
and assuming that aws-region-name is 6 chars for aws-region-name with a
count <= 9.

This commit removes the region name from the healthcheck to give extra
room to have a longer environment name than 9 chars which is currently
needed for the staging-temp environment.

This change could be merged on to master and applied to the production
environment; or alternatively reverted and the limitation on environment
name length documented.