Fix naming of database created within RDS upon creation - Part 2
Specifically for govwifi-admin database.  This pairs with
https://github.com/alphagov/govwifi-terraform/commit/1800868ee465580e50cc5c86c5079f4f596da730
which implements the same fix for two other databases.

The "name" argument passed to aws_db_instance resource creates a
database with that name within the RDS instance.

There is a naming requirement that prevents a hyphen being in the name.

Since an environment can have a hyphen, this commit changes the hyphen
to an underline.

This will likely break the application which accesses the database as it
will be trying to use a hyphen.

This is a temporary fix to allow us to provision the infrastructure for
environment "staging-temp".